+++ src/main/java/com/zy/ssm/dao/ICommentDao.java
@@ -8,5 +8,5 @@ import com.zy.ssm.domain.Comment;
 
 @MapperScan
 public interface ICommentDao {
-	List<Comment> getCommentsByMomentId(Long id);
+	List<Comment> getCommentsByMomentId(Integer id);
 }
